refactor: reorganize package structure for ExamplePresenter and ExampleRepository
