+++ app/src/main/java/com/thechance/qurio/domain/repository/ExampleRepository.kt
@@ -1,4 +1,4 @@
-package com.thechance.qurio.presentation.domain.repository
+package com.thechance.qurio.domain.repository
 
 interface ExampleRepository {
     suspend fun getData(): String
